# packages/contract/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TOMA API",
    "version": "1.0.0",
    "description": "Training management system API. Contract generated from @toma/shared zod schemas."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
...
  "components": {
    "schemas": {
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "default": "about:blank"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          }
        },
        "required": [
          "title",
          "status"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/cycles/board": {
      "get": {
        "tags": [
          "cycles"
        ],
        "summary": "A cycle's bidding/registration board, scoped to the caller (manager bids / HR review)",
        "parameters": [
          {
            "schema": {
              "type": [
                "integer",
                "null"
              ]
            },
            "required": false,
            "name": "cycleId",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "The cycle board",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "cycle": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer",
                          "exclusiveMinimum": 0
                        },
                        "year": {
                          "type": "integer",
                          "minimum": 2000,
                          "maximum": 2100
                        },
                        "quarter": {
                          "type": "integer",
                          "minimum": 1,
                          "maximum": 4
                        },
                        "status": {
                          "type": "string",
                          "enum": [
                            "draft",
                            "bidding",
                            "registration",
                            "locked",
                            "completed"
                          ]
                        },
                        "biddingClosesAt": {
                          "type": [
                            "string",
                            "null"
                          ],
                          "format": "date-time"
                        },
                        "registrationClosesAt": {
                          "type": [
                            "string",
                            "null"
                          ],
                          "format": "date-time"
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      },
                      "required": [
                        "id",
                        "year",
                        "quarter",
                        "status",
                        "biddingClosesAt",
                        "registrationClosesAt",
                        "createdAt"
                      ]
                    },
                    "courses": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "courseId": {
                            "type": "integer",
                            "exclusiveMinimum": 0
                          },
                          "title": {
                            "type": "string"
                          },
                          "discipline": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "lifecycleState": {
                            "type": "string"
                          },
                          "capacity": {
                            "type": [
                              "integer",
                              "null"
                            ]
                          },
                          "totalBidSeats": {
                            "type": "integer",
                            "minimum": 0
                          },
                          "registeredCount": {
                            "type": "integer",
                            "minimum": 0
                          },
                          "waitlistedCount": {
                            "type": "integer",
                            "minimum": 0
                          },
                          "myBidSeats": {
                            "type": [
                              "integer",
                              "null"
                            ],
                            "minimum": 0
                          }
                        },
                        "required": [
                          "courseId",
                          "title",
                          "discipline",
                          "lifecycleState",
                          "capacity",
                          "totalBidSeats",
                          "registeredCount",
                          "waitlistedCount",
                          "myBidSeats"
                        ]
                      }
                    }
                  },
                  "required": [
                    "cycle",
                    "courses"
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Not permitted for this role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/cycles": {
      "post": {
        "tags": [
          "cycles"
        ],
        "summary": "Create a quarterly training cycle (HR)",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "year": {
                    "type": "integer",
                    "minimum": 2000,
                    "maximum": 2100
                  },
                  "quarter": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 4
                  }
                },
                "required": [
                  "year",
                  "quarter"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created cycle",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "exclusiveMinimum": 0
                    },
                    "year": {
                      "type": "integer",
                      "minimum": 2000,
                      "maximum": 2100
                    },
                    "quarter": {
                      "type": "integer",
                      "minimum": 1,
                      "maximum": 4
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "draft",
                        "bidding",
                        "registration",
                        "locked",
                        "completed"
                      ]
                    },
                    "biddingClosesAt": {
                      "type": [
                        "string",
                        "null"
                      ],
                      "format": "date-time"
                    },
                    "registrationClosesAt": {
                      "type": [
                        "string",
                        "null"
                      ],
                      "format": "date-time"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "required": [
                    "id",
                    "year",
                    "quarter",
                    "status",
                    "biddingClosesAt",
                    "registrationClosesAt",
                    "createdAt"
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Not permitted for this role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/cycles/{id}/open-bidding": {
      "post": {
        "tags": [
          "cycles"
        ],
        "summary": "HR opens bidding on candidate courses with a deadline; managers are mailed (req. #1)",
        "parameters": [
          {
            "schema": {
              "type": [
                "integer",
                "null"
              ]
            },
            "required": false,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "biddingClosesAt": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "courseIds": {
                    "type": "array",
                    "items": {
                      "type": "integer",
                      "exclusiveMinimum": 0
                    },
                    "minItems": 1
                  }
                },
                "required": [
                  "biddingClosesAt",
                  "courseIds"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated cycle",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "exclusiveMinimum": 0
                    },
                    "year": {
                      "type": "integer",
                      "minimum": 2000,
                      "maximum": 2100
                    },
                    "quarter": {
                      "type": "integer",
                      "minimum": 1,
                      "maximum": 4
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "draft",
                        "bidding",
                        "registration",
                        "locked",
                        "completed"
                      ]
                    },
                    "biddingClosesAt": {
                      "type": [
                        "string",
                        "null"
                      ],
                      "format": "date-time"
                    },
                    "registrationClosesAt": {
                      "type": [
                        "string",
                        "null"
                      ],
                      "format": "date-time"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "required": [
                    "id",
                    "year",
                    "quarter",
                    "status",
                    "biddingClosesAt",
                    "registrationClosesAt",
                    "createdAt"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/cycles/{id}/open-registration": {
      "post": {
        "tags": [
          "cycles"
        ],
        "summary": "HR opens registration for chosen courses with a lock deadline; managers mailed (#2)",
        "parameters": [
          {
            "schema": {
              "type": [
                "integer",
                "null"
              ]
            },
            "required": false,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "registrationClosesAt": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "courseIds": {
                    "type": "array",
                    "items": {
                      "type": "integer",
                      "exclusiveMinimum": 0
                    },
                    "minItems": 1
                  }
                },
                "required": [
                  "registrationClosesAt",
                  "courseIds"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated cycle",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "exclusiveMinimum": 0
                    },
                    "year": {
                      "type": "integer",
                      "minimum": 2000,
                      "maximum": 2100
                    },
                    "quarter": {
                      "type": "integer",
                      "minimum": 1,
                      "maximum": 4
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "draft",
                        "bidding",
                        "registration",
                        "locked",
                        "completed"
                      ]
                    },
                    "biddingClosesAt": {
                      "type": [
                        "string",
                        "null"
                      ],
                      "format": "date-time"
                    },
                    "registrationClosesAt": {
                      "type": [
                        "string",
                        "null"
                      ],
                      "format": "date-time"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "required": [
                    "id",
                    "year",
                    "quarter",
                    "status",
                    "biddingClosesAt",
                    "registrationClosesAt",
                    "createdAt"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/cycles/{id}/lock": {
      "post": {
        "tags": [
          "cycles"
        ],
        "summary": "Lock registration for a cycle — only HR can change registrations afterwards (#4)",
        "parameters": [
          {
            "schema": {
              "type": [
                "integer",
                "null"
              ]
            },
            "required": false,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Locked cycle",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "exclusiveMinimum": 0
                    },
                    "year": {
                      "type": "integer",
                      "minimum": 2000,
                      "maximum": 2100
                    },
                    "quarter": {
                      "type": "integer",
                      "minimum": 1,
                      "maximum": 4
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "draft",
                        "bidding",
                        "registration",
                        "locked",
                        "completed"
                      ]
                    },
                    "biddingClosesAt": {
                      "type": [
                        "string",
                        "null"
                      ],
                      "format": "date-time"
                    },
                    "registrationClosesAt": {
                      "type": [
                        "string",
                        "null"
                      ],
                      "format": "date-time"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "required": [
                    "id",
                    "year",
                    "quarter",
                    "status",
                    "biddingClosesAt",
                    "registrationClosesAt",
                    "createdAt"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/courses/{id}/bid": {
      "post": {
        "tags": [
          "cycles"
        ],
        "summary": "Set a manager's bid (seats wanted) on a candidate course (requirement #1)",
        "parameters": [
          {
            "schema": {
              "type": [
                "integer",
                "null"
              ]
            },
            "required": false,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "seats": {
                    "type": "integer",
                    "minimum": 0
                  }
                },
                "required": [
                  "seats"
                ]
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Bid saved"
          },
          "403": {
            "description": "Bidding closed / not permitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/courses/{id}/bids": {
      "get": {
        "tags": [
          "cycles"
        ],
        "summary": "All managers’ bids on a candidate course (HR review, requirement #2)",
        "parameters": [
          {
            "schema": {
              "type": [
                "integer",
                "null"
              ]
            },
            "required": false,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Bids",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "courseId": {
                        "type": "integer",
                        "exclusiveMinimum": 0
                      },
                      "managerId": {
                        "type": "string",
                        "minLength": 1
                      },
                      "managerName": {
                        "type": "string"
                      },
                      "seats": {
                        "type": "integer",
                        "minimum": 0
                      },
                      "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                      }
                    },
                    "required": [
                      "courseId",
                      "managerId",
                      "managerName",
                      "seats",
                      "updatedAt"
                    ]
                  }
                }
              }
            }
          }
        }
      }
    },
    "/courses/{id}/decision": {
      "post": {
        "tags": [
          "cycles"
        ],
        "summary": "HR confirms or cancels a course by participant numbers; participants mailed (#6/#7/#8)",
        "parameters": [
          {
            "schema": {
              "type": [
                "integer",
                "null"
              ]
            },
            "required": false,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "decision": {
                    "type": "string",
                    "enum": [
                      "confirm",
                      "cancel"
                    ]
                  }
                },
                "required": [
                  "decision"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "New lifecycle state",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "state": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "state"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}